Separated styles from home screen and replaced hard coded values
